* good synthetic data - FINAL

+++ utils/numpy_utils.py
@@ -6,7 +6,7 @@ class NumpyUtils(object):
 
     @staticmethod
     def random_number_in_range(range_min, range_max):
-        return range_min + np.random.random() * (range_max - range_min)
+        return np.float64(range_min + np.random.random() * (range_max - range_min))
 
     @staticmethod
     def get_vector_size(vec):
